deduplicate privatization approaches
