+++ python/src/opendp/network/__init__.py
@@ -1,2 +1,1 @@
 from .layers import *
-from .odometer import PrivacyOdometer
