+++ python/src/opendp/network/layers/__init__.py
@@ -1,4 +1,3 @@
 from .attention import DPMultiheadAttention
 from .bahdanau import DPBahdanauAttention
 from .lstm import DPLSTM, DPLSTMCell
-from .base import InstanceGrad
