+++ python/src/opendp/network/layers/bahdanau.py
@@ -6,7 +6,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.parameter import Parameter
 
-from opendp.network.layers.base import InstanceGrad
+from opendp.network.odometer_reconstruction import InstanceGrad
 
 
 class DPBahdanauAttentionScale(nn.Module, InstanceGrad):
